# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"BasicResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"object": {}
				},
				"required": [
					"message",
					"object"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EmprestimoRequestsDto": {
				"properties": {
					"idLivro": {
						"type": "number",
						"format": "double"
					},
					"usuarioID": {
						"type": "number",
						"format": "double"
					},
					"dataEmprestimo": {
						"type": "string"
					},
					"dataDevolucao": {
						"type": "string"
					}
				},
				"required": [
					"idLivro",
					"usuarioID",
					"dataEmprestimo",
					"dataDevolucao"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmprestimoDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"idLivro": {
						"type": "number",
						"format": "double"
					},
					"usuarioID": {
						"type": "number",
						"format": "double"
					},
					"dataEmprestimo": {
						"type": "string"
					},
					"dataDevolucao": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"idLivro",
					"usuarioID",
					"dataEmprestimo",
					"dataDevolucao"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoriaRequestsDto": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoriaDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "API Documentação",
		"version": "1.0.0",
		"description": "Documentação da API usando Swagger"
	},
	"paths": {
...
		"/livro/name/{name}": {
			"get": {
				"operationId": "FiltrarCategoriaPorNome",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Livro"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/emprestimo": {
			"post": {
				"operationId": "CadastrarEmprestimo",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmprestimoRequestsDto"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "AtualizarEmprestimo",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmprestimoDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletarEmprestimo",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmprestimoDto"
							}
						}
					}
				}
			}
		},
		"/emprestimo/id/{id}": {
			"get": {
				"operationId": "FiltrarEmprestimoPorId",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/emprestimo/livro/{idLivro}": {
			"get": {
				"operationId": "FiltrarEmprestimoPorLivro",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idLivro",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/emprestimo/usuario/{usuarioID}": {
			"get": {
				"operationId": "FiltrarEmprestimoPorUsuario",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "usuarioID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/emprestimo/all": {
			"get": {
				"operationId": "ListarTodosEmprestimos",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": []
			}
		},
		"/categoria": {
			"post": {
				"operationId": "CadastrarCategoria",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categoria"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoriaRequestsDto"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "AtualizarCategoria",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categoria"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoriaDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletarCategoria",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categoria"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoriaDto"
							}
						}
					}
				}
			}
		},
		"/categoria/id/{id}": {
			"get": {
				"operationId": "FiltrarCategoria",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categoria"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/categoria/name": {
			"get": {
				"operationId": "FiltrarCategoriaPorNome",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categoria"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/categoria/all": {
			"get": {
				"operationId": "ListarTodosCategoria",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categoria"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3040/api",
			"description": "Servidor de Desenvolvimento"
		}
	]
}